Fix for: Typing fast in Manage Video Widget title input field results in error 500
When multiple characters are typed into the text input quickly, that is, faster than the debounce, then this results in multiple save actions being batched into a single request. In this case, exception 'The submitForm() method is being called, but the FormView has already been built. Are you calling $this->getForm() - which creates the FormView - before submitting the form?' is being thrown.